chore:refactor functions to be common
